+++ .vscode-test.mjs
@@ -2,6 +2,6 @@ import { defineConfig } from "@vscode/test-cli";
 
 export default defineConfig({
   files: "dist/test/**/*.test.js",
-  version: "1.101.0",
+  version: "1.93.1",
   workspaceFolder: ".",
 });