+++ src/extension.ts
@@ -1,21 +1,50 @@
 import * as vscode from "vscode";
 import { SUPPORTED_LANGS, SUPPORTED_MODULES } from "./config";
-import ModulesRenameProvider from "./providers/renameProvider";
+import {
+  ModulesRenameProvider,
+  ScriptsRenameProvider,
+} from "./providers/renameProvider";
 import CompletionItemProvider from "./providers/completionProvider";
-import DefinitionProvider from "./providers/definitionProvider";
+import {
+  ScriptDefinitionProvider,
+  ModuleDefinitionProvider as ModuleDefinitionProvider,
+} from "./providers/definitionProvider";
+import Cache from "./libs/cache";
+import loadCaches from "./libs/loadCaches";
 import checkDocument from "./libs/checkDocument";
-import getAllFiles from "./utils/getAllFiles";
+import ClassNameCache from "./libs/classNameCache";
+import { getAllScriptFiles } from "./utils/getAllFiles";
 import CssModuleDependencyCache from "./libs/cssModuleDependencyCache";
 
 export function activate(context: vscode.ExtensionContext) {
   const diagnosticCollection =
     vscode.languages.createDiagnosticCollection("cssModules");
 
+  Cache.context = context;
+  loadCaches(diagnosticCollection);
+
+  vscode.workspace.onDidOpenTextDocument((document) =>
+    checkDocument(document, diagnosticCollection)
+  );
+  vscode.workspace.onDidChangeTextDocument((e) =>
+    checkDocument(e.document, diagnosticCollection)
+  );
+  getAllScriptFiles().then((files) =>
+    files.forEach(async (uri) =>
+      checkDocument(
+        await vscode.workspace.openTextDocument(uri),
+        diagnosticCollection
+      )
+    )
+  );
+
+  // Commands
   const resetCacheCommand = vscode.commands.registerCommand(
     "css-scss-modules-intellisense.resetCache",
     async () => {
       try {
         await CssModuleDependencyCache.populateCacheFromWorkspace();
+        await ClassNameCache.populateCacheFromWorkspace();
         vscode.window.showInformationMessage("Cache has been reset");
       } catch (e) {
         vscode.window.showErrorMessage(e as string);
@@ -33,46 +62,31 @@ export function activate(context: vscode.ExtensionContext) {
   );
 
   // Definition Provider
-  const definitionProvider = vscode.languages.registerDefinitionProvider(
+  const scriptDefinitionProvider = vscode.languages.registerDefinitionProvider(
     SUPPORTED_LANGS,
-    new DefinitionProvider()
+    new ScriptDefinitionProvider()
+  );
+  const moduleDefinitionProvider = vscode.languages.registerDefinitionProvider(
+    SUPPORTED_MODULES,
+    new ModuleDefinitionProvider()
   );
 
   // RenameProvider
-  const renameProvider = vscode.languages.registerRenameProvider(
+  const scriptsRenameProvider = vscode.languages.registerRenameProvider(
+    SUPPORTED_LANGS,
+    new ScriptsRenameProvider()
+  );
+  const modulesRenameProvider = vscode.languages.registerRenameProvider(
     SUPPORTED_MODULES,
     new ModulesRenameProvider()
-  );
-
-  vscode.workspace.onDidOpenTextDocument((document) =>
-    checkDocument(document, diagnosticCollection)
-  );
-  vscode.workspace.onDidChangeTextDocument((e) =>
-    checkDocument(e.document, diagnosticCollection)
-  );
-  getAllFiles().then((files) =>
-    files.forEach(async (uri) =>
-      checkDocument(
-        await vscode.workspace.openTextDocument(uri),
-        diagnosticCollection
-      )
-    )
-  );
-
-  CssModuleDependencyCache.initialize(context);
-  vscode.workspace.onDidCreateFiles((e) => {
-    e.files.forEach((uri) =>
-      CssModuleDependencyCache.updateCacheForDocument({ uri })
-    );
-  });
-  vscode.workspace.onDidChangeTextDocument((e) =>
-    CssModuleDependencyCache.updateCacheForDocument({ document: e.document })
   );
 
   context.subscriptions.push(
     resetCacheCommand,
     completionProvider,
-    definitionProvider,
-    renameProvider
+    moduleDefinitionProvider,
+    scriptDefinitionProvider,
+    modulesRenameProvider,
+    scriptsRenameProvider
   );
 }