+++ src/providers/completionProvider.ts
@@ -1,9 +1,10 @@
 import * as vscode from "vscode";
-import * as fs from "fs";
+import { MESSAGES } from "../config";
+import ClassNameCache from "../libs/classNameCache";
 import isPositionInString from "../utils/isPositionInString";
 import isPositionInComment from "../utils/isPositionInComment";
-import extractClassNames from "../utils/extractClassNames";
-import { resolveImportPathWithAliases } from "../utils/getPath";
+import { getWorkspaceRelativeImportPath } from "../utils/getPath";
+import { getModuleFileRegex } from "../utils/getFileExtensionRegex";
 
 export default class CompletionItemProvider
   implements vscode.CompletionItemProvider
@@ -34,28 +35,27 @@ export default class CompletionItemProvider
 
     const [_, varName] = match;
     const importRegex = new RegExp(
-      `import\\s+${varName}\\s+from\\s+['"](.+\\.module\\.(css|scss|less))['"]`
+      `import\\s+${varName}\\s+from\\s+['"](.+\\.module\\.(${getModuleFileRegex()}))['"]`
     );
     const fullText = document.getText();
-    const imp = fullText.match(importRegex);
-    if (!imp) {
+    const importMatch = fullText.match(importRegex);
+    if (!importMatch) {
       return;
     }
 
-    const resolvedPath = resolveImportPathWithAliases(document, imp[1]);
-
-    if (!fs.existsSync(resolvedPath)) {
+    const classNames = await ClassNameCache.getClassNamesFromImportPath(
+      getWorkspaceRelativeImportPath(document, importMatch[1])
+    );
+    if (!classNames) {
       return;
     }
-
-    const classNames = await extractClassNames(resolvedPath);
     return new vscode.CompletionList(
       classNames.map((name) => {
         const item = new vscode.CompletionItem(
           name,
           vscode.CompletionItemKind.Variable
         );
-        item.detail = "CSS Module class";
+        item.detail = MESSAGES.COMPLETION.CSS_MODULE_CLASS;
         return item;
       }),
       false