+++ src/utils/getAllFiles.ts
@@ -1,10 +1,14 @@
 import * as vscode from "vscode";
+import {
+  getModuleFileRegex,
+  getScriptFileRegex,
+} from "./getFileExtensionRegex";
 
 const config = vscode.workspace.getConfiguration("cssModulesIntellisense");
 const blacklistPatterns = config.get<string[]>("blacklistPatterns", []);
 
-const getAllFiles = async () => {
-  const includePattern = "**/*.{ts,tsx,js,jsx}";
+export const getAllScriptFiles = async () => {
+  const includePattern = `**/*.{${getScriptFileRegex()}}`;
 
   const excludePattern = `{${blacklistPatterns.join(",")}}`;
 
@@ -16,4 +20,15 @@ const getAllFiles = async () => {
   return files;
 };
 
-export default getAllFiles;
+export const getAllModuleFiles = async () => {
+  const includePattern = `**/*.{${getModuleFileRegex()}}`;
+
+  const excludePattern = `{${blacklistPatterns.join(",")}}`;
+
+  const files = await vscode.workspace.findFiles(
+    includePattern,
+    excludePattern
+  );
+
+  return files;
+};